Move all main implementation into app package for all binaries (#7106)
* Move all main implementation into app package for all binaries

Because in Go you cannot import the main package, it makes it difficult
in integration tests to build from source the dapr binaries from the
previous version. With this change, in v1.14 we can import the app
package and build the binaries from source for v1.13.

PR updates binaries to have a minimal main package that calls into the
app package. PR has no functional changes.


* Move options and components to original package location


* Updates copyright to 2023


---------
