+++ cmd/operator/main.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2021 The Dapr Authors
+Copyright 2023 The Dapr Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -14,61 +14,9 @@ limitations under the License.
 package main
 
 import (
-	"github.com/dapr/dapr/cmd/operator/options"
-	"github.com/dapr/dapr/pkg/buildinfo"
-	"github.com/dapr/dapr/pkg/metrics"
-	"github.com/dapr/dapr/pkg/operator"
-	"github.com/dapr/dapr/pkg/operator/monitoring"
-	"github.com/dapr/kit/concurrency"
-	"github.com/dapr/kit/logger"
-	"github.com/dapr/kit/signals"
+	"github.com/dapr/dapr/cmd/operator/app"
 )
 
-var log = logger.NewLogger("dapr.operator")
-
 func main() {
-	opts := options.New()
-
-	// Apply options to all loggers.
-	if err := logger.ApplyOptionsToLoggers(&opts.Logger); err != nil {
-		log.Fatal(err)
-	}
-
-	log.Infof("Starting Dapr Operator -- version %s -- commit %s", buildinfo.Version(), buildinfo.Commit())
-	log.Infof("Log level set to: %s", opts.Logger.OutputLevel)
-
-	metricsExporter := metrics.NewExporterWithOptions(log, metrics.DefaultMetricNamespace, opts.Metrics)
-
-	if err := monitoring.InitMetrics(); err != nil {
-		log.Fatal(err)
-	}
-
-	ctx := signals.Context()
-	op, err := operator.NewOperator(ctx, operator.Options{
-		Config:                              opts.Config,
-		TrustAnchorsFile:                    opts.TrustAnchorsFile,
-		LeaderElection:                      !opts.DisableLeaderElection,
-		WatchdogMaxRestartsPerMin:           opts.MaxPodRestartsPerMinute,
-		WatchNamespace:                      opts.WatchNamespace,
-		ServiceReconcilerEnabled:            !opts.DisableServiceReconciler,
-		ArgoRolloutServiceReconcilerEnabled: opts.EnableArgoRolloutServiceReconciler,
-		WatchdogEnabled:                     opts.WatchdogEnabled,
-		WatchdogInterval:                    opts.WatchdogInterval,
-		WatchdogCanPatchPodLabels:           opts.WatchdogCanPatchPodLabels,
-		APIPort:                             opts.APIPort,
-		HealthzPort:                         opts.HealthzPort,
-	})
-	if err != nil {
-		log.Fatalf("error creating operator: %v", err)
-	}
-
-	err = concurrency.NewRunnerManager(
-		metricsExporter.Run,
-		op.Run,
-	).Run(ctx)
-	if err != nil {
-		log.Fatalf("error running operator: %v", err)
-	}
-
-	log.Info("operator shut down gracefully")
+	app.Run()
 }
